fix: AuthProvider
