--- src/auth/contexts.tsx
@@ -1,11 +1,0 @@
-import { createContext } from "react";
-
-interface AuthContextType {
-  loggedIn: boolean;
-  logIn: () => void;
-  logOut: () => void;
-}
-
-const AuthContext = createContext<AuthContextType>({} as AuthContextType);
-
-export default AuthContext;
