--- src/auth/hook.tsx
@@ -1,7 +1,0 @@
-import { useContext } from "react";
-
-import authContext from "./contexts";
-
-const useAuth = () => useContext(authContext);
-
-export default useAuth;
