+++ src/routes/AppRoutes.tsx
@@ -1,7 +1,6 @@
 import { Routes, Route } from "react-router-dom";
 import { Provider as ReduxProvider } from "react-redux";
 import { ToastContainer, Bounce } from "react-toastify";
-import AuthProvider from "../auth/AuthProvider";
 import "react-toastify/dist/ReactToastify.css";
 import store from "../API/RTKQuery/index";
 import CatalogPage from "../pages/CatalogPage";
@@ -12,50 +11,48 @@ import PrivateRoute from "../auth/PrivateRoute";
 
 const AppRoutes = () => {
   return (
-    <AuthProvider>
-      <ReduxProvider store={store}>
-        <ToastContainer
-          position="bottom-right"
-          autoClose={5000}
-          hideProgressBar={false}
-          newestOnTop={false}
-          closeOnClick
-          rtl={false}
-          pauseOnFocusLoss
-          draggable
-          pauseOnHover
-          theme="colored"
-          transition={Bounce}
+    <ReduxProvider store={store}>
+      <ToastContainer
+        position="bottom-right"
+        autoClose={5000}
+        hideProgressBar={false}
+        newestOnTop={false}
+        closeOnClick
+        rtl={false}
+        pauseOnFocusLoss
+        draggable
+        pauseOnHover
+        theme="colored"
+        transition={Bounce}
+      />
+      <Routes>
+        <Route path="/" element={<Login />} />
+        <Route
+          path="/catalog"
+          element={
+            <PrivateRoute>
+              <CatalogPage />
+            </PrivateRoute>
+          }
         />
-        <Routes>
-          <Route path="/" element={<Login />} />
-          <Route
-            path="/catalog"
-            element={
-              <PrivateRoute>
-                <CatalogPage />
-              </PrivateRoute>
-            }
-          />
-          <Route
-            path="details"
-            element={
-              <PrivateRoute>
-                <HelpRequestPage />
-              </PrivateRoute>
-            }
-          />
-          <Route
-            path="/profile"
-            element={
-              <PrivateRoute>
-                <ProfilePage />
-              </PrivateRoute>
-            }
-          />
-        </Routes>
-      </ReduxProvider>
-    </AuthProvider>
+        <Route
+          path="details"
+          element={
+            <PrivateRoute>
+              <HelpRequestPage />
+            </PrivateRoute>
+          }
+        />
+        <Route
+          path="/profile"
+          element={
+            <PrivateRoute>
+              <ProfilePage />
+            </PrivateRoute>
+          }
+        />
+      </Routes>
+    </ReduxProvider>
   );
 };
